Add files via upload
ignore locked jobs

+++ DeleteOldBuilds.groovy
@@ -1,4 +1,7 @@
 Jenkins.instance.getItemByFullName('jobFullName').builds.findAll { 
-  it.number > 10 && it.number < 6140 }.each {
-   it.delete() 
+  it.number > 0 && it.number < 6140 }.each {
+  //locked job
+    if(it.isKeepLog() != true){
+      it.delete()
+    }
 }
